Move `view-filter` -> `util`
