+++ src/sidebar/services/root-thread.js
@@ -3,6 +3,7 @@ import * as metadata from '../util/annotation-metadata';
 import memoize from '../util/memoize';
 import { generateFacetedFilter } from '../util/search-filter';
 import * as tabs from '../util/tabs';
+import viewFilter from '../util/view-filter';
 
 // Mapping from sort order name to a less-than predicate
 // function for comparing annotations to determine their sort order.
@@ -31,7 +32,7 @@ const sortFns = {
  * The root thread is then displayed by viewer.html
  */
 // @ngInject
-export default function RootThread(annotationsService, store, viewFilter) {
+export default function RootThread(annotationsService, store) {
   /**
    * Build the root conversation thread from the given UI state.
    *
@@ -59,7 +60,7 @@ export default function RootThread(annotationsService, store, viewFilter) {
       });
 
       options.filterFn = function (annot) {
-        return viewFilter.filter([annot], filters).length > 0;
+        return viewFilter([annot], filters).length > 0;
       };
     }
